Backend/requirements.txt
autopep8
celery>=5.0.5
Django>=3.2.9
cloudinary>=1.25.0
django-celery-email>=3.0.0
django-cloudinary-storage>=0.3.0
django-cors-headers>=3.7.0
django-filter>=2.4.0
django-heroku>=0.3.1
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.7.1
djoser>=2.1.0
flake8>=3.9.2
gunicorn>=20.1.0
Pillow>=8.2.0
python-decouple>=3.4
requests>=2.25.1
sentry-sdk>=1.4.3
whitenoise>=5.2.0
django-phonenumber-field
phonenumbers
drf-yasg
redis
unidecode
channels
channels_redis